MVP: Submit expenses to budget via mobile
This commit represents the MVP (minimum viable product) for submitting
expenses to the budget spreadsheet via a mobile device. It removes the
need to retain receipts, as they can be entered in a matter of seconds
from anywhere.

It shares the visual language of the existing spreadsheet.

Also:
- Fails gracefully and reports error while retaining form data
- Allows multiple submissions without reloading

Trello: https://trello.com/c/pCfTGnqW